refactor: 통합테스트와 구분을 위한 unit 패키지 추가
